+++ ui-design-lib/src/main/java/co/electriccoin/zcash/ui/design/component/ZashiTopAppBar.kt
@@ -1,0 +1,34 @@
+package co.electriccoin.zcash.ui.design.component
+
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import co.electriccoin.zcash.ui.design.theme.ZcashTheme
+import co.electriccoin.zcash.ui.design.theme.internal.SecondaryTypography
+import co.electriccoin.zcash.ui.design.theme.internal.TopAppBarColors
+
+@Composable
+@Suppress("LongParameterList")
+fun ZashiSmallTopAppBar(
+    title: String,
+    subtitle: String?,
+    modifier: Modifier = Modifier,
+    showTitleLogo: Boolean = false,
+    colors: TopAppBarColors = ZcashTheme.colors.topAppBarColors,
+    navigationAction: @Composable () -> Unit = {},
+    hamburgerMenuActions: (@Composable RowScope.() -> Unit)? = null,
+    regularActions: (@Composable RowScope.() -> Unit)? = null,
+) {
+    SmallTopAppBar(
+        modifier = modifier,
+        colors = colors,
+        hamburgerMenuActions = hamburgerMenuActions,
+        navigationAction = navigationAction,
+        regularActions = regularActions,
+        subTitle = subtitle,
+        showTitleLogo = showTitleLogo,
+        titleText = title,
+        titleStyle = SecondaryTypography.headlineSmall.copy(fontWeight = FontWeight.SemiBold)
+    )
+}