+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/settings/SettingsViewTestSetup.kt
@@ -3,13 +3,20 @@ package co.electriccoin.zcash.ui.screen.settings
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
-import co.electriccoin.zcash.ui.screen.settings.model.TroubleshootingParameters
+import co.electriccoin.zcash.ui.design.util.stringRes
+import co.electriccoin.zcash.ui.screen.settings.model.SettingsState
+import co.electriccoin.zcash.ui.screen.settings.model.SettingsTroubleshootingState
+import co.electriccoin.zcash.ui.screen.settings.model.TroubleshootingItemState
 import co.electriccoin.zcash.ui.screen.settings.view.Settings
 import java.util.concurrent.atomic.AtomicInteger
 
 class SettingsViewTestSetup(
     private val composeTestRule: ComposeContentTestRule,
-    private val troubleshootingParameters: TroubleshootingParameters
+    isTroubleshootingEnabled: Boolean = false,
+    isBackgroundSyncEnabled: Boolean = false,
+    isKeepScreenOnDuringSyncEnabled: Boolean = false,
+    isAnalyticsEnabled: Boolean = false,
+    isRescanEnabled: Boolean = false
 ) {
     private val onBackCount = AtomicInteger(0)
     private val onFeedbackCount = AtomicInteger(0)
@@ -19,6 +26,30 @@ class SettingsViewTestSetup(
     private val onBackgroundSyncChangedCount = AtomicInteger(0)
     private val onKeepScreenOnChangedCount = AtomicInteger(0)
     private val onAnalyticsChangedCount = AtomicInteger(0)
+
+    private val settingsTroubleshootingState =
+        if (isTroubleshootingEnabled) {
+            SettingsTroubleshootingState(
+                rescan =
+                    TroubleshootingItemState(isRescanEnabled) {
+                        onRescanCount.incrementAndGet()
+                    },
+                backgroundSync =
+                    TroubleshootingItemState(isBackgroundSyncEnabled) {
+                        onBackgroundSyncChangedCount.incrementAndGet()
+                    },
+                keepScreenOnDuringSync =
+                    TroubleshootingItemState(isKeepScreenOnDuringSyncEnabled) {
+                        onKeepScreenOnChangedCount.incrementAndGet()
+                    },
+                analytics =
+                    TroubleshootingItemState(isAnalyticsEnabled) {
+                        onAnalyticsChangedCount.incrementAndGet()
+                    }
+            )
+        } else {
+            null
+        }
 
     fun getBackCount(): Int {
         composeTestRule.waitForIdle()
@@ -64,31 +95,24 @@ class SettingsViewTestSetup(
         composeTestRule.setContent {
             ZcashTheme {
                 Settings(
-                    troubleshootingParameters = troubleshootingParameters,
-                    onBack = {
-                        onBackCount.incrementAndGet()
-                    },
-                    onFeedback = {
-                        onFeedbackCount.incrementAndGet()
-                    },
-                    onAdvancedSettings = {
-                        onAdvancedSettingsCount.incrementAndGet()
-                    },
-                    onAbout = {
-                        onAboutCount.incrementAndGet()
-                    },
-                    onRescanWallet = {
-                        onRescanCount.incrementAndGet()
-                    },
-                    onBackgroundSyncSettingsChanged = {
-                        onBackgroundSyncChangedCount.incrementAndGet()
-                    },
-                    onKeepScreenOnDuringSyncSettingsChanged = {
-                        onKeepScreenOnChangedCount.incrementAndGet()
-                    },
-                    onAnalyticsSettingsChanged = {
-                        onAnalyticsChangedCount.incrementAndGet()
-                    },
+                    state =
+                        SettingsState(
+                            isLoading = false,
+                            version = stringRes("app_version"),
+                            settingsTroubleshootingState = settingsTroubleshootingState,
+                            onBack = {
+                                onBackCount.incrementAndGet()
+                            },
+                            onSendUsFeedbackClick = {
+                                onFeedbackCount.incrementAndGet()
+                            },
+                            onAdvancedSettingsClick = {
+                                onAdvancedSettingsCount.incrementAndGet()
+                            },
+                            onAboutUsClick = {
+                                onAboutCount.incrementAndGet()
+                            },
+                        ),
                     topAppBarSubTitleState = TopAppBarSubTitleState.None,
                 )
             }