+++ ui-lib/src/main/java/co/electriccoin/zcash/di/RepositoryModule.kt
@@ -1,5 +1,7 @@
 package co.electriccoin.zcash.di
 
+import co.electriccoin.zcash.ui.common.repository.ConfigurationRepository
+import co.electriccoin.zcash.ui.common.repository.ConfigurationRepositoryImpl
 import co.electriccoin.zcash.ui.common.repository.WalletRepository
 import co.electriccoin.zcash.ui.common.repository.WalletRepositoryImpl
 import org.koin.core.module.dsl.singleOf
@@ -9,4 +11,5 @@ import org.koin.dsl.module
 val repositoryModule =
     module {
         singleOf(::WalletRepositoryImpl) bind WalletRepository::class
+        singleOf(::ConfigurationRepositoryImpl) bind ConfigurationRepository::class
     }