+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/advancedsettings/AndroidAdvancedSettings.kt
@@ -4,60 +4,51 @@ package co.electriccoin.zcash.ui.screen.advancedsettings
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import co.electriccoin.zcash.di.koinActivityViewModel
-import co.electriccoin.zcash.ui.MainActivity
-import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
+import co.electriccoin.zcash.ui.common.compose.LocalNavController
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
 import co.electriccoin.zcash.ui.screen.advancedsettings.view.AdvancedSettings
+import co.electriccoin.zcash.ui.screen.advancedsettings.viewmodel.AdvancedSettingsViewModel
+import org.koin.androidx.compose.koinViewModel
 
 @Suppress("LongParameterList")
 @Composable
-internal fun MainActivity.WrapAdvancedSettings(
-    goBack: () -> Unit,
+internal fun WrapAdvancedSettings(
     goDeleteWallet: () -> Unit,
     goExportPrivateData: () -> Unit,
-    goChooseServer: () -> Unit,
     goSeedRecovery: () -> Unit,
-    onCurrencyConversion: () -> Unit
 ) {
+    val navController = LocalNavController.current
     val walletViewModel = koinActivityViewModel<WalletViewModel>()
-
+    val viewModel = koinViewModel<AdvancedSettingsViewModel>()
     val walletState = walletViewModel.walletStateInformation.collectAsStateWithLifecycle().value
+    val state =
+        viewModel.state.collectAsStateWithLifecycle().value.copy(
+            onDeleteZashiClick = goDeleteWallet,
+            onExportPrivateDataClick = goExportPrivateData,
+            onRecoveryPhraseClick = goSeedRecovery
+        )
 
-    WrapAdvancedSettings(
-        goBack = goBack,
-        goDeleteWallet = goDeleteWallet,
-        goExportPrivateData = goExportPrivateData,
-        goChooseServer = goChooseServer,
-        goSeedRecovery = goSeedRecovery,
-        topAppBarSubTitleState = walletState,
-        onCurrencyConversion = onCurrencyConversion
-    )
-}
-
-@Composable
-@Suppress("LongParameterList")
-private fun WrapAdvancedSettings(
-    goBack: () -> Unit,
-    goExportPrivateData: () -> Unit,
-    goChooseServer: () -> Unit,
-    goSeedRecovery: () -> Unit,
-    goDeleteWallet: () -> Unit,
-    onCurrencyConversion: () -> Unit,
-    topAppBarSubTitleState: TopAppBarSubTitleState,
-) {
     BackHandler {
-        goBack()
+        viewModel.onBack()
+    }
+
+    LaunchedEffect(Unit) {
+        viewModel.navigationCommand.collect {
+            navController.navigate(it)
+        }
+    }
+
+    LaunchedEffect(Unit) {
+        viewModel.backNavigationCommand.collect {
+            navController.popBackStack()
+        }
     }
 
     AdvancedSettings(
-        onBack = goBack,
-        onDeleteWallet = goDeleteWallet,
-        onExportPrivateData = goExportPrivateData,
-        onChooseServer = goChooseServer,
-        onSeedRecovery = goSeedRecovery,
-        topAppBarSubTitleState = topAppBarSubTitleState,
-        onCurrencyConversion = onCurrencyConversion
+        state = state,
+        topAppBarSubTitleState = walletState,
     )
 }