+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/exchangerate/ZashiButton.kt
@@ -1,21 +1,25 @@
 package co.electriccoin.zcash.ui.screen.exchangerate
 
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Immutable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import co.electriccoin.zcash.ui.design.R
+import co.electriccoin.zcash.ui.design.component.BlankSurface
 import co.electriccoin.zcash.ui.design.component.ButtonState
 import co.electriccoin.zcash.ui.design.component.LottieProgress
+import co.electriccoin.zcash.ui.design.newcomponent.PreviewScreens
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
 import co.electriccoin.zcash.ui.design.util.getValue
 
@@ -23,7 +27,7 @@ import co.electriccoin.zcash.ui.design.util.getValue
 internal fun ZashiButton(
     state: ButtonState,
     modifier: Modifier = Modifier,
-    colors: ButtonColors = ZashiButtonDefaults.primaryButtonColors(),
+    colors: ZashiButtonColors = ZashiButtonDefaults.primaryButtonColors(),
     content: @Composable RowScope.(ZashiButtonScope) -> Unit = ZashiButtonDefaults.content
 ) {
     ZashiButton(
@@ -45,7 +49,7 @@ internal fun ZashiButton(
     modifier: Modifier = Modifier,
     enabled: Boolean = true,
     isLoading: Boolean = false,
-    colors: ButtonColors = ZashiButtonDefaults.primaryButtonColors(),
+    colors: ZashiButtonColors = ZashiButtonDefaults.primaryButtonColors(),
     content: @Composable RowScope.(ZashiButtonScope) -> Unit = ZashiButtonDefaults.content
 ) {
     val scope =
@@ -75,7 +79,8 @@ internal fun ZashiButton(
         modifier = modifier,
         shape = RoundedCornerShape(12.dp),
         enabled = enabled,
-        colors = colors,
+        colors = colors.toButtonColors(),
+        border = colors.borderColor.takeIf { it != Color.Unspecified }?.let { BorderStroke(1.dp, it) },
         content = {
             content(scope)
         }
@@ -104,13 +109,13 @@ object ZashiButtonDefaults {
         contentColor: Color = ZcashTheme.zashiColors.btnPrimaryFg,
         disabledContainerColor: Color = ZcashTheme.zashiColors.btnPrimaryBgDisabled,
         disabledContentColor: Color = ZcashTheme.zashiColors.btnPrimaryFgDisabled,
-    ): ButtonColors =
-        ButtonDefaults.buttonColors(
-            containerColor = containerColor,
-            contentColor = contentColor,
-            disabledContainerColor = disabledContainerColor,
-            disabledContentColor = disabledContentColor
-        )
+    ) = ZashiButtonColors(
+        containerColor = containerColor,
+        contentColor = contentColor,
+        disabledContainerColor = disabledContainerColor,
+        disabledContentColor = disabledContentColor,
+        borderColor = Color.Unspecified
+    )
 
     @Composable
     fun tertiaryButtonColors(
@@ -118,11 +123,88 @@ object ZashiButtonDefaults {
         contentColor: Color = ZcashTheme.zashiColors.btnTertiaryFg,
         disabledContainerColor: Color = Color.Unspecified,
         disabledContentColor: Color = Color.Unspecified,
-    ): ButtonColors =
-        ButtonDefaults.buttonColors(
-            containerColor = containerColor,
-            contentColor = contentColor,
-            disabledContainerColor = disabledContainerColor,
-            disabledContentColor = disabledContentColor
-        )
+    ) = ZashiButtonColors(
+        containerColor = containerColor,
+        contentColor = contentColor,
+        disabledContainerColor = disabledContainerColor,
+        disabledContentColor = disabledContentColor,
+        borderColor = Color.Unspecified
+    )
+
+    @Composable
+    fun destroyButtonColors(
+        containerColor: Color = ZcashTheme.zashiColors.btnDestroyBg,
+        contentColor: Color = ZcashTheme.zashiColors.btnDestroyFg,
+        borderColor: Color = ZcashTheme.zashiColors.btnDestroyBorder,
+        disabledContainerColor: Color = Color.Unspecified,
+        disabledContentColor: Color = Color.Unspecified,
+    ) = ZashiButtonColors(
+        containerColor = containerColor,
+        contentColor = contentColor,
+        disabledContainerColor = disabledContainerColor,
+        disabledContentColor = disabledContentColor,
+        borderColor = borderColor
+    )
 }
+
+@Immutable
+data class ZashiButtonColors(
+    val containerColor: Color,
+    val contentColor: Color,
+    val disabledContainerColor: Color,
+    val disabledContentColor: Color,
+    val borderColor: Color,
+)
+
+@Composable
+private fun ZashiButtonColors.toButtonColors() =
+    ButtonDefaults.buttonColors(
+        containerColor = containerColor,
+        contentColor = contentColor,
+        disabledContainerColor = disabledContainerColor,
+        disabledContentColor = disabledContentColor,
+    )
+
+@Suppress("UnusedPrivateMember")
+@PreviewScreens
+@Composable
+private fun PrimaryPreview() =
+    ZcashTheme {
+        BlankSurface {
+            ZashiButton(
+                modifier = Modifier.fillMaxWidth(),
+                text = "Primary",
+                onClick = {},
+            )
+        }
+    }
+
+@Suppress("UnusedPrivateMember")
+@PreviewScreens
+@Composable
+private fun TertiaryPreview() =
+    ZcashTheme {
+        BlankSurface {
+            ZashiButton(
+                modifier = Modifier.fillMaxWidth(),
+                text = "Primary",
+                colors = ZashiButtonDefaults.tertiaryButtonColors(),
+                onClick = {},
+            )
+        }
+    }
+
+@Suppress("UnusedPrivateMember")
+@PreviewScreens
+@Composable
+private fun DestroyPreview() =
+    ZcashTheme {
+        BlankSurface {
+            ZashiButton(
+                modifier = Modifier.fillMaxWidth(),
+                text = "Primary",
+                colors = ZashiButtonDefaults.destroyButtonColors(),
+                onClick = {},
+            )
+        }
+    }